[ADD] partner_base_multicompany_default
