+++ partner_base_multicompany_default/__init__.py
@@ -1,0 +1,1 @@
+from . import models
