+++ partner_base_multicompany_default/models/__init__.py
@@ -1,0 +1,1 @@
+from . import partner
