+++ partner_base_multicompany_default/readme/CONFIGURE.rst
@@ -1,0 +1,4 @@
+To configure this module, you need to:
+
+#. Install some module that lets you manage partners, such as contact, sale, purchase...
+   (if not already installed).
