+++ partner_base_multicompany_default/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Jairo Llopis (`Moduon <https://www.moduon.team/>`__)
+* Yannick Payot <yannick.payot@camptocamp.com>
