+++ partner_base_multicompany_default/readme/DESCRIPTION.rst
@@ -1,0 +1,12 @@
+On itself this module does nothing.
+
+This module prepare the extention of partner to propagate multi-company
+property field such as partner accounts, fiscal position, delivery carrier from the
+current company to all other companies where you have access rights.
+
+This module is a dependency of:
+
+* partner_account_multicompany_default
+* partner_delivery_multicompany_default
+* partner_product_multicompany_default
+* partner_purchase_multicompany_default
