+++ partner_base_multicompany_default/readme/USAGE.rst
@@ -1,0 +1,5 @@
+To use this module, you need to:
+
+#. Go to a partner's form view.
+#. Enter the "Accounting" tab.
+#. Click on "Propagate receivable/payable account to other companies".
