+++ setup/partner_base_multicompany_default/odoo/addons/partner_base_multicompany_default
@@ -1,0 +1,1 @@
+../../../../partner_base_multicompany_default
